+++ packages/ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vercel/examples-ui",
-  "version": "0.2.7-beta.1",
+  "version": "0.2.7-beta.2",
   "main": "index.tsx",
   "license": "MIT",
   "scripts": {},
